backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
python-multipart>=0.0.9
python-dotenv>=1.0
pydantic>=2.6
pydantic-settings>=2.1
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
pandas>=2.1
openpyxl>=3.1
email-validator>=2.1
rq>=1.15.0
rq-scheduler>=0.13.0
redis>=5.0.0
pytest>=7.4
pytest-asyncio>=0.21
httpx>=0.25
faker>=20.0
pytest-cov>=4.1
sendgrid>=6.11
twilio>=8.10
jinja2>=3.1
